backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "certifi>=2025.8.3",
    "fastapi[standard]>=0.116.1",
    "google-adk>=1.14.0",
    "google-cloud-storage>=2.19.0",
    "lxml>=6.0.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
]
